+++ src/helpers/index.ts
@@ -3,7 +3,6 @@ export * from './error-handler'
 export * from './event-bus'
 export * from './promise.helpers'
 export * from './ethereum.helpers'
-export * from './solana.helpers'
 export * from './formatting.helpers'
 export * from './date-formatting.helpers'
 export * from './tokens.helpers'
