+++ src/types/api.types.ts
@@ -1,0 +1,1 @@
+export type PageOrder = 'asc' | 'desc'
