--- src/types/api/index.ts
@@ -1,6 +1,0 @@
-export * from './default.types'
-export * from './pricer.types'
-export * from './book.types'
-export * from './generator.types'
-export * from './promocode.types'
-export * from './network.types'
