+++ src/types/index.ts
@@ -1,11 +1,14 @@
 export * from './common.types'
 export * from './provider.types'
 export * from './ethereum.types'
-export * from './solana.types'
 export * from './contracts'
-export * from './api'
 export * from './chain.types'
 export * from './guide.types'
 export * from './symbols'
 export * from './nft-tokens.types'
 export * from './marketplace.types'
+export * from './api.types'
+export * from './book.types'
+export * from './generator.types'
+export * from './pricer.types'
+export * from './promocode.types'
